requirements.txt
flask
requests
beautifulsoup4
dnspython
selenium
shodan
PyQt6
matplotlib
pyvis
reportlab
